API Types Cleanup (#11238)
