+++ services/app-api/handlers/banners/create.test.ts
@@ -1,8 +1,10 @@
 import { createBanner } from "./create";
 import { APIGatewayProxyEvent } from "aws-lambda";
-import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
+// types
+import { StatusCodes } from "../../utils/types";
+// utils
 import { error } from "../../utils/constants/constants";
+import { proxyEvent } from "../../utils/testing/proxyEvent";
 
 jest.mock("../../utils/auth/authorization", () => ({
   isAuthorized: jest.fn().mockReturnValue(true),
