+++ services/app-api/handlers/banners/create.ts
@@ -1,10 +1,12 @@
 import * as yup from "yup";
 import handler from "../handler-lib";
+// utils
 import dynamoDb from "../../utils/dynamo/dynamodb-lib";
 import { hasPermissions } from "../../utils/auth/authorization";
 import { validateData } from "../../utils/validation/validation";
 import { error } from "../../utils/constants/constants";
-import { StatusCodes, UserRoles } from "../../utils/types/types";
+// types
+import { StatusCodes, UserRoles } from "../../utils/types";
 
 export const createBanner = handler(async (event, _context) => {
   if (!hasPermissions(event, [UserRoles.ADMIN])) {
