+++ services/app-api/handlers/banners/delete.test.ts
@@ -1,8 +1,10 @@
 import { deleteBanner } from "./delete";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
 import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
 import { error } from "../../utils/constants/constants";
+// types
+import { StatusCodes } from "../../utils/types/other";
 
 jest.mock("../../utils/auth/authorization", () => ({
   isAuthorized: jest.fn().mockReturnValue(true),
