+++ services/app-api/handlers/banners/fetch.test.ts
@@ -1,12 +1,14 @@
 import { fetchBanner } from "./fetch";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
 import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
 import { error } from "../../utils/constants/constants";
 import {
   mockBannerResponse,
   mockDocumentClient,
 } from "../../utils/testing/setupJest";
+// types
+import { StatusCodes } from "../../utils/types";
 
 jest.mock("../../utils/auth/authorization", () => ({
   isAuthorized: jest.fn().mockReturnValue(true),
