+++ services/app-api/handlers/handler-lib.ts
@@ -1,13 +1,15 @@
-import * as debug from "../utils/debugging/debug-lib";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
+import * as debug from "../utils/debugging/debug-lib";
 import { isAuthorized } from "../utils/auth/authorization";
 import {
   internalServerError,
   buildResponse,
 } from "../utils/responses/response-lib";
 import { error } from "../utils/constants/constants";
-import { StatusCodes } from "../utils/types/types";
 import { sanitizeObject } from "../utils/sanitize/sanitize";
+// types
+import { StatusCodes } from "../utils/types";
 
 type LambdaFunction = (
   event: APIGatewayProxyEvent, // eslint-disable-line no-unused-vars
