+++ services/app-api/handlers/reports/fetch.test.ts
@@ -1,7 +1,7 @@
 import { fetchReport, fetchReportsByState } from "./fetch";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
 import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
 import { error } from "../../utils/constants/constants";
 import {
   mockDocumentClient,
@@ -10,6 +10,8 @@ import {
   mockReportFieldData,
   mockDynamoDataCompleted,
 } from "../../utils/testing/setupJest";
+// types
+import { StatusCodes } from "../../utils/types";
 
 jest.mock("../../utils/auth/authorization", () => ({
   isAuthorized: jest.fn().mockReturnValue(true),
