+++ services/app-api/handlers/reports/submit.test.ts
@@ -1,7 +1,7 @@
 import { submitReport } from "./submit";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
 import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
 import { error } from "../../utils/constants/constants";
 import {
   mockApiKey,
@@ -10,6 +10,8 @@ import {
   mockDynamoDataCompleted,
   mockDynamoDataMLRComplete,
 } from "../../utils/testing/setupJest";
+// types
+import { StatusCodes } from "../../utils/types/other";
 
 jest.mock("../../utils/auth/authorization", () => ({
   isAuthorized: jest.fn().mockReturnValue(true),
