+++ services/app-api/handlers/reports/update.test.ts
@@ -1,14 +1,16 @@
 import { fetchReport } from "./fetch";
 import { updateReport } from "./update";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
 import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
 import {
   mockDynamoData,
   mockMcparReport,
   mockReportFieldData,
 } from "../../utils/testing/setupJest";
 import { error } from "../../utils/constants/constants";
+// types
+import { StatusCodes } from "../../utils/types";
 
 jest.mock("../../utils/auth/authorization", () => ({
   isAuthorized: jest.fn().mockResolvedValue(true),
