+++ services/app-api/handlers/templates/fetch.test.ts
@@ -1,8 +1,10 @@
 import { fetchTemplate } from "./fetch";
 import { APIGatewayProxyEvent } from "aws-lambda";
+// utils
 import { proxyEvent } from "../../utils/testing/proxyEvent";
-import { StatusCodes } from "../../utils/types/types";
 import { error } from "../../utils/constants/constants";
+// types
+import { StatusCodes } from "../../utils/types";
 
 jest.mock("aws-sdk", () => ({
   S3: jest.fn().mockImplementation(() => ({
