+++ services/app-api/utils/auth/authorization.test.ts
@@ -1,6 +1,6 @@
 import { proxyEvent } from "../testing/proxyEvent";
 import { hasPermissions, isAuthorized } from "./authorization";
-import { UserRoles } from "../types/types";
+import { UserRoles } from "../types/users";
 
 const mockVerifier = jest.fn();
