+++ services/app-api/utils/s3/s3-lib.ts
@@ -1,7 +1,7 @@
 import { Credentials, S3, Endpoint } from "aws-sdk";
 import { ServiceConfigurationOptions } from "aws-sdk/lib/service";
 import { buckets } from "../constants/constants";
-import { S3Copy, S3Get, S3Put, State } from "../types/types";
+import { S3Copy, S3Get, S3Put, State } from "../types/other";
 
 export const createS3Client = () => {
   const s3Config: S3.ClientConfiguration &
