+++ services/app-api/utils/types/banners.ts
@@ -1,0 +1,27 @@
+// BANNER
+
+export interface BannerData {
+  title: string;
+  description: string;
+  link?: string;
+  [key: string]: any;
+}
+
+export interface AdminBannerData extends BannerData {
+  key: string;
+  startDate: number;
+  endDate: number;
+  isActive?: boolean;
+}
+
+export interface AdminBannerMethods {
+  fetchAdminBanner: Function;
+  writeAdminBanner: Function;
+  deleteAdminBanner: Function;
+}
+
+export interface AdminBannerShape extends AdminBannerMethods {
+  bannerData?: AdminBannerData;
+  isLoading: boolean;
+  errorMessage?: string;
+}
