+++ services/app-api/utils/types/reports.ts
@@ -1,0 +1,177 @@
+import { FormJson } from "./formFields";
+import { AnyObject, CompletionData, CustomHtmlElement, State } from "./other";
+
+// REPORT STRUCTURE
+
+export interface ReportJson {
+  id?: string;
+  type?: string;
+  name: string;
+  basePath: string;
+  adminDisabled?: boolean;
+  routes: ReportRoute[];
+  flatRoutes?: ReportRoute[];
+  validationSchema: AnyObject;
+}
+
+export type ReportRoute = ReportRouteWithForm | ReportRouteWithoutForm;
+
+export interface ReportRouteBase {
+  name: string;
+  path: string;
+  pageType?: string;
+}
+
+export type ReportRouteWithForm =
+  | StandardReportPageShape
+  | DrawerReportPageShape
+  | ModalDrawerReportPageShape
+  | ModalOverlayReportPageShape;
+
+export interface ReportPageShapeBase extends ReportRouteBase {
+  children?: never;
+  verbiage: ReportPageVerbiage;
+}
+
+export interface StandardReportPageShape extends ReportPageShapeBase {
+  form: FormJson;
+  dashboard?: never;
+  modalForm?: never;
+  overlayForm?: never;
+  drawerForm?: never;
+  entityType?: never;
+}
+
+export interface DrawerReportPageShape extends ReportPageShapeBase {
+  entityType: string;
+  verbiage: DrawerReportPageVerbiage;
+  drawerForm: FormJson;
+  modalForm?: never;
+  overlayForm?: never;
+  form?: never;
+}
+
+export interface ModalDrawerReportPageShape extends ReportPageShapeBase {
+  entityType: string;
+  verbiage: ModalDrawerReportPageVerbiage;
+  modalForm: FormJson;
+  drawerForm: FormJson;
+  overlayForm?: never;
+  form?: never;
+}
+
+export interface ModalOverlayReportPageShape extends ReportPageShapeBase {
+  entityType: string;
+  verbiage: ModalOverlayReportPageVerbiage;
+  modalForm: FormJson;
+  overlayForm?: FormJson;
+  drawerForm?: never;
+  form?: never;
+}
+
+export interface ReportRouteWithoutForm extends ReportRouteBase {
+  children?: ReportRoute[];
+  pageType?: string;
+  entityType?: never;
+  verbiage?: never;
+  modalForm?: never;
+  overlayForm?: never;
+  drawerForm?: never;
+  form?: never;
+}
+
+export interface ReportPageVerbiage {
+  intro: {
+    section: string;
+    subsection?: string;
+    spreadsheet?: string;
+    info?: string | CustomHtmlElement[];
+  };
+}
+
+export interface DrawerReportPageVerbiage extends ReportPageVerbiage {
+  dashboardTitle: string;
+  countEntitiesInTitle?: boolean;
+  drawerTitle: string;
+  drawerInfo?: CustomHtmlElement[];
+  missingEntityMessage?: CustomHtmlElement[];
+}
+
+export interface ModalDrawerReportPageVerbiage
+  extends DrawerReportPageVerbiage {
+  addEntityButtonText: string;
+  editEntityButtonText: string;
+  addEditModalAddTitle: string;
+  addEditModalEditTitle: string;
+  addEditModalMessage: string;
+  deleteEntityButtonAltText: string;
+  deleteModalTitle: string;
+  deleteModalConfirmButtonText: string;
+  deleteModalWarning: string;
+  entityUnfinishedMessage: string;
+  enterEntityDetailsButtonText: string;
+  editEntityDetailsButtonText: string;
+}
+
+export interface ModalOverlayReportPageVerbiage extends ReportPageVerbiage {
+  addEntityButtonText: string;
+  dashboardTitle: string;
+  countEntitiesInTitle: boolean;
+  tableHeader: string;
+  addEditModalHint: string;
+  emptyDashboardText: string;
+}
+
+// REPORT METADATA
+
+export interface ReportMetadata {
+  archived: boolean;
+  reportType: string;
+  submittedBy?: string;
+  createdAt: number;
+  lastAltered: number;
+  state: State;
+  id: string;
+  submittedOnDate?: string;
+  fieldDataId: string;
+  formTemplateId: string;
+  lastAlteredBy: string;
+  status: string;
+  isComplete: boolean;
+  completionStatus?: CompletionData;
+}
+
+export interface MLRReportMetadata extends ReportMetadata {
+  locked: boolean;
+  reportType: "MLR";
+  submissionName: string;
+  submissionCount: number;
+  previousRevisions: string[];
+}
+
+export interface MCPARReportMetadata extends ReportMetadata {
+  programName: string;
+  reportType: "MCPAR";
+  reportingPeriodStartDate: number;
+  reportingPeriodEndDate: number;
+  dueDate: number;
+  combinedData: boolean;
+}
+
+/**
+ * Type guard to perform run-time checks on report types.
+ *
+ * Use this function on data retrieved from Dynamo allow your data to be safely typed.
+ * @param report any report type
+ * @returns
+ */
+export function isMLRReportMetadata(
+  report: unknown
+): report is MLRReportMetadata {
+  return (
+    (report as MLRReportMetadata).reportType === "MLR" &&
+    (report as MLRReportMetadata).locked !== undefined &&
+    (report as MLRReportMetadata).submissionCount !== undefined &&
+    (report as MLRReportMetadata).previousRevisions !== undefined
+  );
+}
