--- services/app-api/utils/types/types.ts
@@ -1,555 +1,0 @@
-// GLOBAL
-
-export interface AnyObject {
-  [key: string]: any;
-}
-
-export interface DynamoGet {
-  TableName: string;
-  Key: {
-    [key: string]: any;
-  };
-}
-
-export interface DynamoWrite {
-  TableName: string;
-  Item: { [key: string]: any };
-}
-
-export interface DynamoDelete {
-  TableName: string;
-  Key: { [key: string]: any };
-}
-
-export interface DynamoUpdate {
-  TableName: string;
-  Key: {
-    key: string;
-  };
-  UpdateExpression?: string;
-  ExpressionAttributeNames: { [key: string]: string };
-  ExpressionAttributeValues: { [key: string]: any };
-}
-
-export interface DynamoScan {
-  TableName: string;
-  [key: string]: any;
-}
-
-export const enum RequestMethods {
-  POST = "POST",
-  GET = "GET",
-  PUT = "PUT",
-  DELETE = "DELETE",
-}
-
-export const enum StatusCodes {
-  SUCCESS = 200,
-  CREATED = 201,
-  BAD_REQUEST = 400,
-  UNAUTHORIZED = 403,
-  NOT_FOUND = 404,
-  SERVER_ERROR = 500,
-}
-
-export interface S3Put extends S3Get {
-  Body: string;
-  ContentType: string;
-}
-
-export interface S3Get {
-  Bucket: string;
-  Key: string;
-}
-
-export interface S3Copy {
-  Bucket: string;
-  CopySource: string;
-  Key: string;
-}
-
-export interface CompletionData {
-  [key: string]: boolean | CompletionData;
-}
-
-// USERS
-
-export enum UserRoles {
-  ADMIN = "mdctmcr-bor", // "MDCT MCR Business Owner Representative"
-  HELP_DESK = "mdctmcr-help-desk", // "MDCT MCR Help Desk"
-  APPROVER = "mdctmcr-approver", // "MDCT MCR Approver"
-  STATE_REP = "mdctmcr-state-rep", // "MDCT MCR State Representative"
-  STATE_USER = "mdctmcr-state-user", // "MDCT MCR State User"
-}
-
-export interface MCRUser {
-  email: string;
-  given_name: string;
-  family_name: string;
-  full_name: string;
-  state?: string;
-  userRole?: string;
-  userIsAdmin?: boolean;
-  userIsHelpDeskUser?: boolean;
-  userIsApprover?: boolean;
-  userIsStateRep?: boolean;
-  userIsStateUser?: boolean;
-}
-
-export interface UserContextShape {
-  user?: MCRUser;
-  showLocalLogins?: boolean;
-  logout: () => Promise<void>;
-  loginWithIDM: () => void;
-  updateTimeout: () => void;
-  getExpiration: Function;
-}
-
-// REPORT STRUCTURE
-
-export interface ReportJson {
-  id?: string;
-  type?: string;
-  name: string;
-  basePath: string;
-  adminDisabled?: boolean;
-  routes: ReportRoute[];
-  flatRoutes?: ReportRoute[];
-  validationSchema: AnyObject;
-}
-
-export type ReportRoute = ReportRouteWithForm | ReportRouteWithoutForm;
-
-export interface ReportRouteBase {
-  name: string;
-  path: string;
-  pageType?: string;
-}
-
-export type ReportRouteWithForm =
-  | StandardReportPageShape
-  | DrawerReportPageShape
-  | ModalDrawerReportPageShape
-  | ModalOverlayReportPageShape;
-
-export interface ReportPageShapeBase extends ReportRouteBase {
-  children?: never;
-  verbiage: ReportPageVerbiage;
-}
-
-export interface StandardReportPageShape extends ReportPageShapeBase {
-  form: FormJson;
-  dashboard?: never;
-  modalForm?: never;
-  overlayForm?: never;
-  drawerForm?: never;
-  entityType?: never;
-}
-
-export interface DrawerReportPageShape extends ReportPageShapeBase {
-  entityType: string;
-  verbiage: DrawerReportPageVerbiage;
-  drawerForm: FormJson;
-  modalForm?: never;
-  overlayForm?: never;
-  form?: never;
-}
-
-export interface ModalDrawerReportPageShape extends ReportPageShapeBase {
-  entityType: string;
-  verbiage: ModalDrawerReportPageVerbiage;
-  modalForm: FormJson;
-  drawerForm: FormJson;
-  overlayForm?: never;
-  form?: never;
-}
-
-export interface ModalOverlayReportPageShape extends ReportPageShapeBase {
-  entityType: string;
-  verbiage: ModalOverlayReportPageVerbiage;
-  modalForm: FormJson;
-  overlayForm?: FormJson;
-  drawerForm?: never;
-  form?: never;
-}
-
-export interface ReportRouteWithoutForm extends ReportRouteBase {
-  children?: ReportRoute[];
-  pageType?: string;
-  entityType?: never;
-  verbiage?: never;
-  modalForm?: never;
-  overlayForm?: never;
-  drawerForm?: never;
-  form?: never;
-}
-
-export interface ReportPageVerbiage {
-  intro: {
-    section: string;
-    subsection?: string;
-    spreadsheet?: string;
-    info?: string | CustomHtmlElement[];
-  };
-}
-
-export interface DrawerReportPageVerbiage extends ReportPageVerbiage {
-  dashboardTitle: string;
-  countEntitiesInTitle?: boolean;
-  drawerTitle: string;
-  drawerInfo?: CustomHtmlElement[];
-  missingEntityMessage?: CustomHtmlElement[];
-}
-
-export interface ModalDrawerReportPageVerbiage
-  extends DrawerReportPageVerbiage {
-  addEntityButtonText: string;
-  editEntityButtonText: string;
-  addEditModalAddTitle: string;
-  addEditModalEditTitle: string;
-  addEditModalMessage: string;
-  deleteEntityButtonAltText: string;
-  deleteModalTitle: string;
-  deleteModalConfirmButtonText: string;
-  deleteModalWarning: string;
-  entityUnfinishedMessage: string;
-  enterEntityDetailsButtonText: string;
-  editEntityDetailsButtonText: string;
-}
-
-export interface ModalOverlayReportPageVerbiage extends ReportPageVerbiage {
-  addEntityButtonText: string;
-  dashboardTitle: string;
-  countEntitiesInTitle: boolean;
-  tableHeader: string;
-  addEditModalHint: string;
-  emptyDashboardText: string;
-}
-
-// REPORT PROVIDER/CONTEXT
-
-export interface ReportKeys {
-  reportType: string;
-  state: string;
-  id: string;
-}
-
-export interface ReportMetadataShape extends ReportKeys {
-  reportType: string;
-  programName: string;
-  status: ReportStatus;
-  reportingPeriodStartDate: number;
-  reportingPeriodEndDate: number;
-  dueDate: number;
-  createdAt: number;
-  lastAltered: number;
-  lastAlteredBy: string;
-  combinedData: boolean;
-  submittedBy?: string;
-  submitterEmail?: string;
-  submittedOnDate?: number;
-  archived?: boolean;
-}
-
-export interface ReportShape extends ReportMetadataShape {
-  formTemplate: ReportJson;
-  fieldData: AnyObject;
-}
-
-export interface ReportContextMethods {
-  fetchReport: Function;
-  fetchReportsByState: Function;
-  archiveReport: Function;
-  createReport: Function;
-  updateReport: Function;
-  clearReportSelection: Function;
-  setReportSelection: Function;
-}
-
-export interface ReportContextShape extends ReportContextMethods {
-  report: ReportShape | undefined;
-  reportsByState: ReportMetadataShape[] | undefined;
-  errorMessage?: string | undefined;
-  lastSavedTime?: string | undefined;
-}
-
-export enum ReportStatus {
-  NOT_STARTED = "Not started",
-  IN_PROGRESS = "In progress",
-  SUBMITTED = "Submitted",
-}
-
-// FORM & FIELD STRUCTURE
-
-export declare type EntityType =
-  | "plans"
-  | "bssEntities"
-  | "accessMeasures"
-  | "qualityMeasures"
-  | "sanctions";
-
-export enum ModalDrawerEntityTypes {
-  ACCESS_MEASURES = "accessMeasures",
-  QUALITY_MEASURES = "qualityMeasures",
-  SANCTIONS = "sanctions",
-}
-
-export interface EntityShape {
-  id: string;
-  [key: string]: any;
-}
-
-export interface FormJson {
-  id: string;
-  fields: FormField[];
-  options?: AnyObject;
-  validation?: AnyObject;
-  adminDisabled?: boolean;
-}
-
-export interface DependentFieldValidation {
-  type: string;
-  dependentFieldName: string;
-  parentOptionId?: never;
-}
-
-export interface NestedFieldValidation {
-  type: string;
-  nested: true;
-  parentFieldName: string;
-  parentOptionId: string;
-}
-
-export interface NestedDependentFieldValidation {
-  type: string;
-  dependentFieldName: string;
-  nested: true;
-  parentFieldName: string;
-  parentOptionId: string;
-}
-
-export type FieldValidationObject =
-  | DependentFieldValidation
-  | NestedFieldValidation
-  | NestedDependentFieldValidation;
-
-export interface FormField {
-  id: string;
-  type: string;
-  validation: string | FieldValidationObject;
-  hydrate?: string;
-  props?: AnyObject;
-  choices?: FieldChoice[];
-  repeat?: string;
-}
-
-export interface DropdownOptions {
-  label: string;
-  value: string;
-}
-
-export interface FieldChoice {
-  id: string;
-  name: string;
-  label: string;
-  value: string;
-  checked?: boolean;
-  children?: FormField[];
-  checkedChildren?: any;
-}
-
-export interface ChoiceFieldProps {
-  name: string;
-  label: string;
-  choices: FieldChoice[];
-  sxOverride?: AnyObject;
-  [key: string]: any;
-}
-export interface Choice {
-  key: string; // choice.name
-  value: string; // choice.value
-}
-
-export interface DropdownChoice {
-  label: string;
-  value: string;
-}
-
-export enum PageTypes {
-  STANDARD = "standard",
-  DRAWER = "drawer",
-  MODAL_DRAWER = "modalDrawer",
-  REVIEW_SUBMIT = "reviewSubmit",
-}
-
-// BANNER
-
-export interface BannerData {
-  title: string;
-  description: string;
-  link?: string;
-  [key: string]: any;
-}
-
-export interface AdminBannerData extends BannerData {
-  key: string;
-  startDate: number;
-  endDate: number;
-  isActive?: boolean;
-}
-
-export interface AdminBannerMethods {
-  fetchAdminBanner: Function;
-  writeAdminBanner: Function;
-  deleteAdminBanner: Function;
-}
-
-export interface AdminBannerShape extends AdminBannerMethods {
-  bannerData?: AdminBannerData;
-  isLoading: boolean;
-  errorMessage?: string;
-}
-
-// ALERTS
-
-export enum AlertTypes {
-  ERROR = "error",
-  INFO = "info",
-  SUCCESS = "success",
-  WARNING = "warning",
-}
-
-// TIME
-
-export interface DateShape {
-  year: number;
-  month: number;
-  day: number;
-}
-
-export interface TimeShape {
-  hour: number;
-  minute: number;
-  second: number;
-}
-
-// OTHER
-
-export interface CustomHtmlElement {
-  type: string;
-  content: string | any;
-  as?: string;
-  props?: AnyObject;
-}
-
-export const enum TemplateKeys {
-  MCPAR = "templates/MCPAR_Workbook_AddProgramName_2022.xlsx",
-  MLR = "templates/MLR_Workbook_2022.xlsx",
-  NAAAR = "templates/NAAAR_Workbook_2022.xlsx",
-}
-
-export type State =
-  | "AL"
-  | "AK"
-  | "AZ"
-  | "AR"
-  | "CA"
-  | "CO"
-  | "CT"
-  | "DE"
-  | "DC"
-  | "FL"
-  | "GA"
-  | "HI"
-  | "ID"
-  | "IL"
-  | "IN"
-  | "IA"
-  | "KS"
-  | "KY"
-  | "LA"
-  | "ME"
-  | "MD"
-  | "MA"
-  | "MI"
-  | "MN"
-  | "MS"
-  | "MO"
-  | "MT"
-  | "NE"
-  | "NV"
-  | "NH"
-  | "NJ"
-  | "NM"
-  | "NY"
-  | "NC"
-  | "ND"
-  | "OH"
-  | "OK"
-  | "OR"
-  | "PA"
-  | "PR"
-  | "RI"
-  | "SC"
-  | "SD"
-  | "TN"
-  | "TX"
-  | "UT"
-  | "VT"
-  | "VA"
-  | "WA"
-  | "WV"
-  | "WI"
-  | "WY";
-
-export interface ReportMetadata {
-  archived: boolean;
-  reportType: string;
-  submittedBy?: string;
-  createdAt: number;
-  lastAltered: number;
-  state: State;
-  id: string;
-  submittedOnDate?: string;
-  fieldDataId: string;
-  formTemplateId: string;
-  lastAlteredBy: string;
-  status: string;
-  isComplete: boolean;
-  completionStatus?: CompletionData;
-}
-
-export interface MLRReportMetadata extends ReportMetadata {
-  locked: boolean;
-  reportType: "MLR";
-  submissionName: string;
-  submissionCount: number;
-  previousRevisions: string[];
-}
-
-export interface MCPARReportMetadata extends ReportMetadata {
-  programName: string;
-  reportType: "MCPAR";
-  reportingPeriodStartDate: number;
-  reportingPeriodEndDate: number;
-  dueDate: number;
-  combinedData: boolean;
-}
-
-/**
- * Type guard to perform run-time checks on report types.
- *
- * Use this function on data retrieved from Dynamo allow your data to be safely typed.
- * @param report any report type
- * @returns
- */
-export function isMLRReportMetadata(
-  report: unknown
-): report is MLRReportMetadata {
-  return (
-    (report as MLRReportMetadata).reportType === "MLR" &&
-    (report as MLRReportMetadata).locked !== undefined &&
-    (report as MLRReportMetadata).submissionCount !== undefined &&
-    (report as MLRReportMetadata).previousRevisions !== undefined
-  );
-}
