+++ services/app-api/utils/types/users.ts
@@ -1,0 +1,32 @@
+// USERS
+
+export enum UserRoles {
+  ADMIN = "mdctmcr-bor", // "MDCT MCR Business Owner Representative"
+  HELP_DESK = "mdctmcr-help-desk", // "MDCT MCR Help Desk"
+  APPROVER = "mdctmcr-approver", // "MDCT MCR Approver"
+  STATE_REP = "mdctmcr-state-rep", // "MDCT MCR State Representative"
+  STATE_USER = "mdctmcr-state-user", // "MDCT MCR State User"
+}
+
+export interface MCRUser {
+  email: string;
+  given_name: string;
+  family_name: string;
+  full_name: string;
+  state?: string;
+  userRole?: string;
+  userIsAdmin?: boolean;
+  userIsHelpDeskUser?: boolean;
+  userIsApprover?: boolean;
+  userIsStateRep?: boolean;
+  userIsStateUser?: boolean;
+}
+
+export interface UserContextShape {
+  user?: MCRUser;
+  showLocalLogins?: boolean;
+  logout: () => Promise<void>;
+  loginWithIDM: () => void;
+  updateTimeout: () => void;
+  getExpiration: Function;
+}
