+++ services/app-api/utils/validation/completionSchemas.ts
@@ -6,7 +6,7 @@ import {
   string,
   number as yupNumber,
 } from "yup";
-import { Choice } from "../types/types";
+import { Choice } from "../types";
 
 export const error = {
   REQUIRED_GENERIC: "A response is required",
